requirements.txt
fastapi
uvicorn
scikit-learn
numpy
sqlalchemy
pymysql
python-dotenv
sentence-transformers
torch
